Implement Poloniex exchange integration
- Add HTTP client with HMAC-SHA256 signing and rate limiting
- Add WebSocket manager for real-time orderbook subscriptions
- Implement PoloniexExchange with Exchange trait methods
- Add connect/disconnect with clock drift detection
- Add subscribe_orderbook with automatic ping loop
- Add test functions for WebSocket, HTTP client, and exchange

+++ src/config/exchange.rs
@@ -23,7 +23,7 @@ pub struct ExchangeConfig {
     /// Taker fee as a decimal string (e.g., "0.001" for 0.1%).
     pub fee_taker: Option<String>,
     /// Maximum API requests per minute.
-    pub rate_limit: Option<i32>,
+    pub rate_limit: Option<i64>,
     /// WebSocket connection settings.
     pub websocket: Option<WebSocketConfig>,
 }
